requirements.txt
streamlit==1.37.1
yfinance==0.2.52
feedparser==6.0.11
requests==2.32.3
python-dotenv==1.0.1
wikipedia==1.4.0
tqdm==4.66.4
openai==1.43.0
